requirements.txt
fastapi
uvicorn
pydantic
requests
pymupdf
chromadb
sentence-transformers
numpy
python-multipart
pytesseract
pillow
